refactor: Further cleanup of tests and validation
